Tests/LibWeb/Text/expected/XML/xml-fragment-innerhtml.txt

child
ex
https://example.test/ns
cafe
touché
